requirements.txt
opencv-python
opencv-contrib-python
numpy
scipy
pillow
flask
flask-socketio
eventlet
deepface
tf-keras
retina-face
scikit-learn